Correctly find callInvite to cancel
I'm a novice at iOS development, but in my attempt to copy this QuickStart implementation into https://github.com/hoxfon/react-native-twilio-programmable-voice/pull/138 I needed to make this change to get the cancel to work correctly.

I wanted to share this change with you, but I have not tested it in this repo myself. Please confirm and merge if it is an issue.